Make launch parameters more customizable
While building and testing a tool provider, I needed the ability to
customize more of the launch parameters; in particular, I wanted to
change the `roles` of the user. While I was there, I switched some of
the other hardcoded parameters to draw from the tool config form as
well.